consolidate and add error handling
